Set node in org-node-capture-target with proposed title
When capturing via org-protocol, we first set te proposed title and
id.

In this case, check if there is an existing node, otherwise this
function errors trying to open the file or buffer as one or both
already exist.

This stops the org-protocol capture from silently "doing nothing", and
allows you to add notes to an existing captured node if desired.